requirements.txt
mysql-connector-python
peewee
pymysql
python-dateutil
requests
dnspython
